Correctly fetch data from TDT: it was ignoring the data buffer's size change I was doing with ActiveX, so the TDT code now makes the buffer quite large indeed.
Important: previous version was recording the trigger channel (for
debugging purposes), and this revision fixes that.

Fix a bunch of matrix-index-out-of-bounds problems that popped up during
channel averaging due to a bug.

Display: fix filtering, display trends and de-trend each channel
independently, bring ROI right up to the end of stimulation.