Add a function to do just one stimulation sequence and process the data:
function [ data response_detected voltage ] = stimulate(stim, hardware,
handles)

Rather than incrementing or decrementing current on a timer, this will be
called as fast as possible by the (nascent) optimisation code.
Consequently, the data acquisition now can be handled using either the
timer callback or from this optimisation code (not in this commit). This
is untested and probably doesn't quite work yet.