Wrote detail code for individual electrode scaling
In plexme.m, updated some code responsible for updating the figure from
a save file to reflect the fact that a) the negative-first checkboxes
are gone, and b) there are now stim-scale objects in its place.

In stimulate.m, added code that creates a new file for every single
electrode (could potentially be  optimized later by checking for
similarities between each electrode's configuration to reduce the number
of written files?  Can be added later relatively easily if the
experiment tends to use repeated electrode configurations and it
significantly slows down program run-time).  The lines that decide
whether or not to use a negative-first wave in the stimulator were
replaced with lines to simply apply each electrode's own custom stim.
The "plexon_write_rectangular_pulse_file" function was edited to be
backwards compatible, but also to add function for inserting a pre-pulse
delay (which will be used in the code that calls it to change the stim
on an electrode-by-electrode basis).  Finally, added some notes on what
to remove later.

Once again, code was not thoroughly tested due to lack of hardware,
however simple gui management works, and no errors are thrown until a
plexon-related function is called as the timer is starting.