Two problems:
(1) The de-bounce isn't working at all.  It mutes all outputs.

(2) Normalisation.

Taking out the normc makes the frequency of responses much closer to
correct, but still not right.  Those weird +1 and -1 in the network stream
don't make any sense to me.  I'd say that I need to re-work normalisation
and be correct about it.