Main structure for Smoothness Checking
Modified the figure to include the Smoothness scanning function, as well
as some parameters to adjust.  Currently being set up to select a random
active electrode, then scan over its amplitude and prepulse pause
duration to determine if it has a response.

Notes: Check line 2539 for a logic question (asked because of comparison
to similar logic in line 2518)
Want to confirm what a good maximum value is for the prepulse pause
duration.  Currently set at 250us.